refactor: extract useHasPlayAudioAction hook for the player-preview gate
Replace the `withSelect` HOC in `play-audio/check.js` with a `useSelect`
based hook, per AGENTS.md ("function components + hooks only — no class
components"):

- Move the pure `selectHasPlayAudioAction( select )` predicate into
  `play-audio/hooks.js` and add a `useHasPlayAudioAction()` wrapper. Pure
  selector stays the shared seam (callable inside `useSelect`/`withSelect`,
  unit-testable without a render); co-locating it with the hook avoids a
  `check.js` <-> `hooks.js` import cycle.
- `PlayAudioCheck` becomes a function component that calls the hook. Its
  prop-driven named export was unused, so nothing else changes.
- The `beyondwords/player` block uses the hook instead of an inline
  `useSelect`.

Left as-is on purpose:

- `PreviewPanel` keeps its own `hasGeneratedContent` predicate — it is
  intentionally looser than `selectHasPlayAudioAction` (e.g. it hides for
  client-side posts with no content id), so unifying them is a deliberate
  sidebar behaviour change for a separate commit.
- `play-audio/index.js`'s own four-prop `withSelect` (a bigger, separate
  change).
